ChildAwareModelBase relays IsDirty changes
In https://github.com/Catel/Catel/issues/1166 a problem is described
that when a nested structure of ChildAwareModelBase classes is used that
the propagation of the IsDirty flag is only done for one level, and not
all the way to the top. By removing code that specifically prohibited
this, the expected functionality is now available.